Initial systemd+gpiod installation files
These files are meant for use on recent Pi OS like Raspberry Pi OS Buster(Legacy) and Bullseye.

Major changes include using systemd to start/stop the service and adding the installation of the gpiod package to control GPIO.

The version of the service installed by this installation script is based on gpiod instead of the deprecated SysFS GPIO control method.